requirements.txt
flask
pillow
pytesseract
fpdf
